docs: add KDocs for models package
