+++ application/src/main/kotlin/models/SettingsModel.kt
@@ -11,6 +11,13 @@ import viewmodels.graphs.GraphViewModel
 import viewmodels.pages.GraphPageViewModel
 import java.io.File
 
+/**
+ * The SettingsModel class is responsible for managing application settings,
+ * connecting to databases, saving and loading graphs, and validating graph names.
+ *
+ * @property isNeo4jConnected indicates whether a Neo4J database is connected
+ * @property graphNameRegEx a regular expression used to validate graph names
+ */
 class SettingsModel {
 	private val jsonDB = GraphJSONDatabase()
 	private val neo4jDB = Neo4jRepository()
@@ -18,6 +25,13 @@ class SettingsModel {
 	var isNeo4jConnected = false
 	val graphNameRegEx = Regex("[a-zA-Z][a-zA-Z0-9_-]*")
 
+	/**
+	 * Connects to a Neo4J database using the provided URI, username, and password.
+	 *
+	 * @param uri the URI of the Neo4J database
+	 * @param user the username for authentication
+	 * @param password the password for authentication
+	 */
 	fun connectToNeo4J(uri: String, user: String, password: String) {
 		isNeo4jConnected = false
 		try {
@@ -30,6 +44,15 @@ class SettingsModel {
 		}
 	}
 
+	/**
+	 * Saves the graph in the specified format and returns a [GraphInfo] object.
+	 *
+	 * @param graphPageViewModel the [GraphPageViewModel] containing the graph to be saved
+	 * @param savingType the type of database to save the graph in
+	 * @param folderPath the path of the folder where the graph will be saved
+	 *
+	 * @return a [GraphInfo] object containing information about the saved graph, or null if the graph couldn't be saved
+	 */
 	fun saveGraph(
 		graphPageViewModel: GraphPageViewModel,
 		savingType: GraphSavingType,
@@ -71,6 +94,14 @@ class SettingsModel {
 		)
 	}
 
+	/**
+	 * Loads a graph from a SQLite database file and updates the [GraphPageViewModel] accordingly.
+	 *
+	 * @param graphPageViewModel the [GraphPageViewModel] to update with the loaded graph
+	 * @param path the path of the SQLite database file to load the graph from
+	 *
+	 * @return `true` if the graph was successfully loaded, `false` otherwise
+	 */
 	fun loadGraphFromSQLite(graphPageViewModel: GraphPageViewModel, path: String): Boolean {
 		try {
 			val graphViewModel = sqliteDB.loadGraph(path)
@@ -85,6 +116,12 @@ class SettingsModel {
 		return true
 	}
 
+	/**
+	 * Saves the graph in a SQLite database file.
+	 *
+	 * @param graphViewModel the [GraphViewModel] containing the graph to be saved
+	 * @param folderPath the path of the folder where the SQLite database file will be saved
+	 */
 	fun saveGraphBySQLite(graphViewModel: GraphViewModel, folderPath: String) {
 		try {
 			sqliteDB.writeDb(graphViewModel, folderPath)
@@ -93,6 +130,13 @@ class SettingsModel {
 		}
 	}
 
+	/**
+	 * Loads a graph from a Neo4J database and updates the [GraphPageViewModel] accordingly.
+	 *
+	 * @param graphPageViewModel the [GraphPageViewModel] to update with the loaded graph
+	 *
+	 * @return `true` if the graph was successfully loaded, `false` otherwise
+	 */
 	fun loadGraphFromNEO4J(graphPageViewModel: GraphPageViewModel): Boolean {
 		if (!isNeo4jConnected) {
 			println("Can't load graph because not found connection to neo4j DB")
@@ -124,6 +168,11 @@ class SettingsModel {
 		return true
 	}
 
+	/**
+	 * Saves the graph in Neo4J database.
+	 *
+	 * @param graphPageViewModel the [GraphPageViewModel] containing the graph to be saved.
+	 */
 	private fun saveGraphByNeo4j(graphPageViewModel: GraphPageViewModel) {
 		if (!isNeo4jConnected) {
 			println("Can't save graph because not found connection to neo4j DB")
@@ -136,6 +185,12 @@ class SettingsModel {
 		}
 	}
 
+	/**
+	 * Saves the graph in a JSON file.
+	 *
+	 * @param graphViewModel the [GraphViewModel] containing the graph to be saved
+	 * @param folderPath the path of the folder where the JSON file will be saved
+	 */
 	private fun saveGraphByJSON(graphViewModel: GraphViewModel, folderPath: String) {
 		val file = File(folderPath, "${graphViewModel.graph.label}.json")
 		try {
@@ -146,6 +201,11 @@ class SettingsModel {
 	}
 
 	companion object {
+		/**
+		 * Loads the application settings from a configuration file.
+		 *
+		 * @return a [SettingsModel] object containing the loaded settings
+		 */
 		@JvmStatic
 		fun loadSettings(): SettingsModel {
 			// TODO(valid load of settings from configuration)
