+++ application/src/main/kotlin/models/utils/GraphInfo.kt
@@ -6,6 +6,16 @@ import utils.GraphSavingType
 import java.nio.file.Path
 import java.time.LocalDateTime
 
+/**
+ * Represents a graph information item in the application.
+ *
+ * @param name the name of the graph
+ * @property folderPath the path to the folder where the graph is stored
+ * @property savingType the type of graph saving (e.g., Neo4j DB, CSV, etc.)
+ * @property previouslyOpenDateTime the date and time when the graph was last opened
+ * @param isHidden indicates whether the graph info item is hidden
+ * @param onClick a callback function to be invoked when the graph info item is clicked
+ */
 class GraphInfo(
 	name: String,
 	val folderPath: String,
@@ -22,14 +32,11 @@ class GraphInfo(
 	} else {
 		{ println("Click on graph info item $name") }
 	},
-	icon = Icons.Default.AutoGraph
-	// TODO(Choose item icon as random from files)
+	icon = Icons.Default.AutoGraph // TODO(Choose item icon as random from files)
 ), Comparable<GraphInfo> {
 	override fun compareTo(other: GraphInfo): Int {
 		return previouslyOpenDateTime.compareTo(other.previouslyOpenDateTime)
 	}
-
-
 
 	override fun toString(): String {
 		return "GraphInfo(name = $mainText, folder = $folderPath, previouslyOpenDateTime = $previouslyOpenDateTime)"
